fix(*): ui updates
